+++ core/src/main/java/de/otto/jlineup/config/Cookie.java
@@ -1,6 +1,7 @@
 package de.otto.jlineup.config;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import java.util.Date;
@@ -22,7 +23,10 @@ public class Cookie {
 
     public final Boolean showInReport;
 
-    public Cookie(String name, String value, String domain, String path, Date expiry, boolean secure, Boolean showInReport) {
+    @JsonIgnore
+    public final boolean screenshotContextGiving;
+
+    public Cookie(String name, String value, String domain, String path, Date expiry, boolean secure, Boolean showInReport, boolean screenshotContextGiving) {
         this.name = name;
         this.value = value;
         this.domain = domain;
@@ -30,6 +34,7 @@ public class Cookie {
         this.expiry = expiry;
         this.secure = secure;
         this.showInReport = showInReport;
+        this.screenshotContextGiving = screenshotContextGiving;
     }
 
     public Cookie(String name, String value) {
@@ -40,6 +45,7 @@ public class Cookie {
         this.expiry = null;
         this.secure = false;
         this.showInReport = null;
+        this.screenshotContextGiving = false;
     }
 
     private Cookie(Builder builder) {
@@ -50,6 +56,7 @@ public class Cookie {
         expiry = builder.expiry;
         secure = builder.secure;
         showInReport = builder.showInReport;
+        screenshotContextGiving = builder.screenshotContextGiving;
     }
 
     public static Builder cookieBuilder() {
@@ -64,6 +71,8 @@ public class Cookie {
         builder.path = copy.getPath();
         builder.expiry = copy.getExpiry();
         builder.secure = copy.isSecure();
+        builder.showInReport = copy.getShowInReport();
+        builder.screenshotContextGiving = copy.isScreenshotContextGiving();
         return builder;
     }
 
@@ -107,6 +116,10 @@ public class Cookie {
         return showInReport;
     }
 
+    public boolean isScreenshotContextGiving() {
+        return screenshotContextGiving;
+    }
+
     /*
      *
      *
@@ -120,19 +133,6 @@ public class Cookie {
      */
 
     @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Cookie cookie = (Cookie) o;
-        return secure == cookie.secure && Objects.equals(name, cookie.name) && Objects.equals(value, cookie.value) && Objects.equals(domain, cookie.domain) && Objects.equals(path, cookie.path) && Objects.equals(expiry, cookie.expiry) && Objects.equals(showInReport, cookie.showInReport);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name, value, domain, path, expiry, secure, showInReport);
-    }
-
-    @Override
     public String toString() {
         return "Cookie{" +
                 "name='" + name + '\'' +
@@ -142,9 +142,22 @@ public class Cookie {
                 ", expiry=" + expiry +
                 ", secure=" + secure +
                 ", showInReport=" + showInReport +
+                ", screenshotContextGiving=" + screenshotContextGiving +
                 '}';
     }
 
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Cookie cookie = (Cookie) o;
+        return secure == cookie.secure && screenshotContextGiving == cookie.screenshotContextGiving && Objects.equals(name, cookie.name) && Objects.equals(value, cookie.value) && Objects.equals(domain, cookie.domain) && Objects.equals(path, cookie.path) && Objects.equals(expiry, cookie.expiry) && Objects.equals(showInReport, cookie.showInReport);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name, value, domain, path, expiry, secure, showInReport, screenshotContextGiving);
+    }
 
     public static final class Builder {
         private String name;
@@ -154,6 +167,7 @@ public class Cookie {
         private Date expiry;
         private boolean secure;
         public Boolean showInReport;
+        public boolean screenshotContextGiving;
 
         private Builder() {
         }
@@ -192,6 +206,11 @@ public class Cookie {
 
         public Builder withShowInReport(Boolean val){
             showInReport = val;
+            return this;
+        }
+
+        public Builder withScreenshotContextGiving(boolean val){
+            screenshotContextGiving = val;
             return this;
         }